backendModels/capibara6_original/requirements.txt
# Framework web
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Base de datos
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0

# Redis & Background Tasks
redis==5.0.1
rq==1.15.1

# Vector stores & Embeddings
faiss-cpu==1.7.4
chromadb==0.4.18
sentence-transformers==2.2.2

# E2B Execution
e2b-code-interpreter==0.1.0

# Fine-tuning
peft==0.7.0
bitsandbytes==0.41.3

# Monitoreo
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilidades
PyJWT==2.8.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
numpy==1.26.2
scikit-learn==1.3.2

# Coqui TTS - Alta calidad en español (solo para VM, NO Vercel)
TTS>=0.22.0

# Kyutai TTS - Delayed Streams Modeling (nueva integración - reemplaza Coqui)
moshi>=0.2.6
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
transformers>=4.35.0
huggingface-hub>=0.19.0